requirements.txt
certifi==2020.11.8
chardet==3.0.4
charset-normalizer==2.0.4
click==7.1.2
Flask==1.1.2
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
paho-mqtt==1.5.1
Pillow==8.3.2
PyYAML==5.3.1
requests==2.24.0
urllib3==1.25.11
Werkzeug==1.0.1
wget==3.2
xmpppy==0.6.2
